chore(tools): relocate music calibration helpers
